Alright! I saw that I had made a mistake previously, I had extended CLSIS3820ScalerView for my new view class incorporating dark current information, instead of CLSSIS3820ScalerChannelView. Now I think I've correctly implemented a class that provides display information about dark current measurement value and measurement state upon request. The old (mistake) class still exists, but header and class file contents are entirely commented out. Also fixed some issues with dark current dwell time being passed as an integer instead of a double. Builds and runs without error, but the class itself hasn't been instantiated/tested yet.

+++ source/beamline/AMDetectorTriggerSource.cpp
@@ -27,11 +27,11 @@ AMDetectorDwellTimeSource::AMDetectorDwellTimeSource(const QString &name, QObjec
 }
 
 void AMDetectorDwellTimeSource::requestSetDwellTime(double dwellSeconds){
-	emit setDwellTime(dwellSeconds);
+    emit setDwellTime(dwellSeconds);
 }
 
 #include <QDebug>
-void AMDetectorDwellTimeSource:: requestSetDarkCurrentCorrectionTime(int timeSeconds) {
+void AMDetectorDwellTimeSource:: requestSetDarkCurrentCorrectionTime(double timeSeconds) {
     qDebug() << "Want to emit darkCurrentCorrectionTime as " << timeSeconds << " in " << name();
     emit setDarkCurrentCorrectionTime(timeSeconds);
